fix: generate configuration automatically to Reusable Workflows
## Problem

The "Create a release API" in GitHub Releases was found to utilize a "committed" configuration file.

- See: [Create a release API](https://docs.github.com/en/rest/releases/releases?apiVersion=2022-11-28#create-a-release)

This resulted in ignoring locally existing configuration files,
causing Release Reusable Workflows to generate release notes without utilizing a configuration file.

## Expected Behavior

Release notes should be generated using the configuration file created in the workflow.

## Solution

To address this issue, the workaround has been modified with the following process:

1. Generate a configuration file.
2. Create a temporary branch and commit the generated configuration file.
3. Push the temporary branch.
4. Execute "Generate release notes content for a release API" to exclusively generate release notes.
5. Create a release using the release notes generated in Step 4.
6. Delete the temporary branch that is no longer needed.

The crucial step is Step 4.
When executing "Generate release notes content for a release API",
specify the commit SHA of the temporary branch in the body parameter `target_commitish`.
This ensures the generation of release notes using the configuration file produced in the workflow.

- See: [Generate release notes content for a release API](https://docs.github.com/en/rest/releases/releases?apiVersion=2022-11-28#generate-release-notes-content-for-a-release)